requirements.txt
aiohttp==3.8.5
discord.py==2.3.2
psutil==5.9.0
python-dotenv==1.0.1
pytz==2022.7
wavelink==3.2.0
yarl==1.9.2

